actual working slider up to date
